requirements.txt
amqp==2.5.2
asgiref==3.2.5
Babel==2.8.0
billiard==3.6.3.0
celery==4.4.1
Django==3.0.7
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-extensions==3.1.3
django-simple-history==3.0.0
django-timezone-field==4.0
djangorestframework==3.12.1
djangorestframework-datatables==0.6.0
flower==0.9.3
gunicorn==20.0.4
importlib-metadata==1.5.0
install==1.3.3
kombu==4.6.8
numpy==1.19.0
pandas==1.0.5
psycopg2==2.8.4
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2019.3
six==1.15.0
sqlparse==0.3.1
tornado==5.1.1
vine==1.3.0
zipp==3.1.0
